Added the MapImageManager and map thumbnail images.
The MapImageManager creates a thumbnail image of a map in a subdirectory
where the map is located.

The MapsDock displays a thumbnail when a map is selected.

ZomboidScene displays a propertly-scaled thumbnail image of a map
during drag-and-drop.  Also, 'highlight current layer' now highlights
the level rather than a specific layer.

When moving map objects representing Lots with ObjectSelectionTool, the
Lot tiles are hidden and the MapObjectItem displays the semi-opaque map
thumbnail image.