Experimental 'sparse' TileLayer to save much memory.
Project Zomboid maps can be as large as 300x300 with over 100 tile layers.
Except for the ground floor layer, most layers have much less than 90,000
tiles assigned.  This implementation replaces the QVector<> with a QHash<>
to track non-empty cells.

Loading all maps in the maptools folder used to take ~700MB of ram, now
it takes ~200MB.

As a side benefit, it is now fast to check whether an entire tile layer is
empty, although that could perhaps have been accomplished by maintaining a
count of non-empty cells as they are assigned.