Added the MiniMap.
The MiniMap is a child widget of MapView.  It displays an up-to-date image
of the map being edited.  It can be resized using buttons in the MiniMap,
and hidden/shown using the menu command or 'M' hotkey.

MapDocument::regionAltered is a new signal emitted whenever the contents of
a TileLayer change.  Unlike regionEdited, regionAltered is emitted during
undo/redo, not just when the user edits something.

CompositeLayerGroup maintains a separate 'visible' flag for each layer
separate from Layer::mVisible.  This is to support multiple MapComposites
using the same Map.