Fixed switch-layer button in TilesetDock not appearing in the extension menu.
When there isn't enough room to display a toolbar button, the toolbar's
'extension' toolbutton appears which displays the remaining actions in a
popup menu.  But if you call QToolbar::addWidget(), that action won't
appear in the popup when the toolbar's parent isn't a QMainWindow.

So, the old QToolButton was replaced by a QAction.

Qt docs says:
When a QToolBar is not a child of a QMainWindow, it loses the ability
to populate the extension pop up with widgets added to the toolbar using
addWidget(). Please use widget actions created by inheriting QWidgetAction
and implementing QWidgetAction::createWidget() instead.