Fixed redrawing the whole MiniMap when using the StampBrush to draw stamps.
When using the StampBrush to draw lines (by Shift-clicking), the BrushItem's
tile layer is the size of the entire map (see StampBrush::configureBrush).

PaintTileLayer would set its mPaintedRegion to that total area (see the
constructor), resulting in PaintTileLayer::undo/redo causing the whole
MiniMap to be redrawn when the regionAltered signal was emitted.

Now I pass BrushItem::tileRegion() as a mask to PaintTileLayer, so it knows
exactly which parts of the map were actually changed.