Fixed selection in tileset list not matching the hidden tab bar's current index.
Tiled's tab bar wasn't removed - only hidden - when the tileset list was
added to the TilesetDock.  The list selection is supposed to stay in synch
with the tab bar's current index.

When adding a new tileset, the list's selection was cleared (due to the
TilesetThumbModel calling reset()).

When removing a tileset that wasn't the current one (via undo/redo), the
list's selection was cleared.  It was ok when deleting the current tileset,
because the tab bar's current index would change.

This could result in a crash because the move-up/move-down buttons were
synched with the tab bar's current index, while the moveTilesetUp()/
moveTilesetDown() methods acted on the (out-of-date) mThumbView index.