Read and write building config files in the user's home directory, etc.
BuildingTiles.txt etc are now saved in ~/.TileZed rather than the application
directory.  This was changed to avoid the user's changes to these files being
lost when installing a new version of TileZed.

MapBaseXMLLots.txt was replaced by TMXConfig.txt, a SimpleFile that lists the
tilesets and names of layers that go in generated .tmx files.