Layer opacity slider affects only a single layer, not an entire level.
In the original Tiled, the opacity slider affected a single layer.
When isometric rendering was changed from layer-by-layer to cell-by-cell,
the opacity slider was changed to affect an entire level. Now the original
behavior is restored.  This required MapComposite storing the opacity for
each tile layer separately from the underlying Map.

SaveAsImageDialog got a new option "Force 100% layer opacity".