Fixed wrong wall tile being chosen when a window or door had no tile assigned.
If the universal "none" tile entry was assigned to a door, and the building was
saved and loaded, the door object would end up with tile=0 and not the
"none" tile entry.  This broke the getWallOffset() method which checks for a
non-zero tile to determine if a door or window exists.