Added Wall tool to BuildingEd.
This tool is for adding or replacing walls.  The big office building in
suburbs1 has intermixed glass/concrete wall tiles as well as internal rooms
that can't be handled by the Pencil tool.

An additional change: "None" is now a valid choice for exterior and interior
walls.  By assigning "None" to a WallObject a section of wall can be removed.