Added "Used Tiles" and "Used Furniture" categories.
Buildings have lists of used tiles and used furniture.  As objects are added
or their tiles are changed, the tiles are added to the appropriate list.
Tiles are never removed from these lists.

Building templates have lists of used tiles and used furniture.  When creating
a template from a building, the building's "used" lists are copied to the
new template.  There is currently no way to edit these lists.

FurnitureTiles gained a pointer to the FurnitureGroup it belongs to.  This
is to support the "Used Furniture" list.

BuildingTileCategory gained a NoneBuildingTileEntry.  This is to support the
"Used Tiles" list.

FurnitureView draws the FurnitureGroup label.  This is to support the "Used Furniture"
list.