Added floor/wall grime tiles to building rooms.
New tile categories "Grime - Floors" and "Grime - Walls" were added.
Each room can be assigned a tile from the new grime categories.

For walls, the correct tile for walls/windows/doors is placed in the
WallGrime layer.

For floors, grime is placed at the base of west walls in the FloorGrime
layer and at the base of north walls in the FloorGrime2 layer.  (Two
layers are needed because there is no proper north-west-corner floor grime
tile.) No floor grime is place in front of doors or on east/south walls.