Allow PathGenerators to draw tiles at an offset from a given Path.
To better support multiple generators per path, a generator can get a new
set of points that is offset from the original path.  The PathStroke class
is based on some work I did with MonkVG.

Also, Path points may be at the center of tiles or between tiles.  This is
needed to support odd-width paths.