Allow map loading when tileset images couldn't be found.
When loading a map, if a tileset image couldn't be found/read, a fake tileset
image is created of the appropriate size.

PathGeneratorMgr uses "missing" tilesets as well so the program can start
even when the Tiles directory hasn't been chosen yet.  When the Tiles directory
is changed (in the PathGeneratorTilesDialog) an attempt is made to reload
the missing tileset images.