BuildingEd uses a single "missing" tile for missing tileset images.
When a tileset can't be found we still need to set the tileset's source()
to something valid so the path can be written properly to TMXConfig.txt.

When displaying the tiles in a missing tileset, we show nothing rather
than the single ??? tile.

When resolving a BuildingTile to a Tiled::Tile, return an actual tile
from a missing tileset rather than the TilesetManager's missingTile().
Otherwise the various category views display any tiles from missing tilesets
under the name "missing" rather than the actual tileset names.