Rewrote TilesetDock for fast map switching.
The TilesetDock used to display a row of tabs, one for each tileset.
Each tab had its own TilesetView and TilesetModel.
Every time the current document changed, the tabs and views were recreated.

Even after I added a list of tileset names, the tabs were still used but
were hidden, resulting in a lot of fiddly code and the same slow performance.

Now the tabs are gone.  There is only a single TilesetView and TilesetModel.