PG_WallGrime remembers which wall tiles it is covering.
PG_WallGrime needs to know which wall tiles it should apply grime to.
Currently it is using BuildingEd's exterior_walls and interior_walls entries.
But the map .tmx file doesn't save those entries, meaning the result of the
generator could change when BuildingEd's entries change.
So now PG_WallGrime saves up to 4 sets of wall tiles it applied grime to.
Once the .tmx is saved and reloaded, the generator won't apply grime to any
tiles except the ones it knew about.