Give nicer resize behavior to BuildingTilesDialog.
The QSplitter between the tileset names list and tile view was replaced
with a horizontal layout, and the tileset names list is just given a fixed
width based on the maximum width of tileset names.

When resizing the window, extra space is given to the widgets on the left
rather than being distributed to both sides (is this "nicer"?).

Also added a horizontal line above the buttons along the bottom.