Fixed display issues with user-placed tiles.
If there are no user-placed tiles in any layer on a floor then the layer-group
has empty bounds.  In order for the tool-tiles to show up, the layer-group
bounds must be force to be non-empty.

Whenever the current layer changes, I force that layer to be non-empty by
calling CompositeLayerGroup::setLayerNonEmpty.