Hack to force CompositeLayerGroup margins to maximum.
TileLayer::drawMargins() returns the area around a map cell needed to contain
the largest tile drawn in that layer. When no tiles have been drawn in a
TileLayer, that layer's draw-margins are all zero.

When in BuildingEd's tile-edit mode, the DrawTileTool wants to display some
tiles that are being drawn with.  This only works if the layer's draw-margins
aren't zero.  Rather than setting the draw-margins to non-zero when
DrawTileTool becomes active, I force them to be non-zero all the time.

This change is in addition to the other hack involving setLayerNonEmpty()
which forces a tile layer to be considered non-empty when a layer is the
current one being edited.

Rather than having every Tileset track maximum tile size and tileset offset
it would be simpler to have the map do it when tilesets are added.