Added BaseFloorEditor, BuildingRenderer classes, etc.
Adding support for both orthogonal and isometric building views.
The goal is to allow the building tools to work the same in both orientations.

BuildingRenderer is analogous to Tiled::MapRenderer; this class renders
building objects and handles coordinate transformations.

BuildingObject gained calcShape() to return a polygon (in tile coords) that
is used to render the object.