Added ShadowBuilding class to better handle live updates during editing.
The old BuildingMap::setCursorObject implementation worked for a single
object, but we need to support live updates while moving multpile objects
around.

The ShadowBuilding class keeps a copy of the building being edited and applies
BuildingModifiers to it.  BuildingModifiers are used to add objects (like
cursor objects) and to move/resize existing objects.