Improvments to the tile properties dialog.
Added copying and pasting tile properties.  Multiple tiles may be copied
from and pasted to.  So the properties of a whole set of wall tiles may be
copied and then pasted onto another set of wall tiles.

The properties of selected tiles may be reset to the defaults.

Added a toolbar with copy/paste/reset buttons.

Added a label that shows the offset of the tile the mouse pointer is over
relative to a selected tile.

Added the Edit menu.

Save/restore the window geometry.
Save/restore the directory the last .tiles file was opened from.

Use QVariant for getting/setting property values.
Created PropGenericEnum base class for all enum properties.