Removed all the "special" enum properites from TileProperties.txt handling.
The WallStyle, DoorStyle, RoofStyle etc properties used to have subclasses
of their own, but it seems that isn't needed.  PropGenericEnum was changed
to handle the different enum properties.

UIProperty::ChangeProperty() no longer updates TileDefTile::mProperties,
removing the need to keep a reference to it.  Now mProperties is only updated
when reading/writing the .tiles file.

Also improved error handling reading TileProperties.txt.

The background color of tilesis changed to indicate any unknown properties.

The tile tooltip shows the properties that will be written to the .tiles file
as well as any unknown properties.