Added tileset history to TileDefDialog, etc.
This history feature is similar to a web browser's go forward/back buttons.

Added a toolbar.  Buttons everywhere!

Save/restore the tile scale.

Fixed the wrong tileset being deleted, and the dialog's list of tilesets
not matching the order of the TileDefFile tilesets.

Try to maintain which row is selected when resetting the list of tilesets.

Fixed crash with undo/redo when the affected tile wasn't in the displayed
tileset.