Added "modify" commands to TileProperties.txt.
Any number of "modify" blocks can be added to TileProperties.txt.
Each "modify" block contains a series of commands that are run against
the properties dictionary of each tile after being read in from a .tiles
file.  These modifier commands can edit the properties of a tile.
This can be used to remove/rename/replace properties and values.