The relative path and extension of lot maps is saved, not just the name.
Since the first days of TileZed, map objects representing lot maps stored the
name of the lot .tmx file in the object type field.  This did not include either
the path to the .tmx file nor the file extension/suffix.  This behavior led to
a number of ass-faced issues in both TileZed and WorldEd.

When a "lot" object was found, the lot .tmx was searched for relative to
the map's file, then inside whatever directory was displayed in the Maps
dock.  Simply browsing to another directory in the Maps dock would cause
the lot manager to attempt to reload every lot.

When reading a .tmx file, the absolute path to a lot-map is now resolved
relative to the .tmx file and is stored in a map-object's type field.
When saving a .tmx file the relative path to lot-maps is saved, including
the extension.

The tooltip for map-object items shows the relative path to a lot-map.
The Objects dock shows the relative path to a lot-map, but double-clicking
to edit the filename presents the absolute path.

The *only* reason I didn't fix this long ago is because I didn't want to
break any of lemmy's tools that I didn't have access to back then.