BuildingMap adds TilesetManager::missingTile() to maps.
If a building tile couldn't be found, the special "missing" tile is added
to the map representing the building.  In order to track whether a map has
any missing tilesets we need to add TilesetManager::mMissingTileset to the
building's map.

This was more important in WorldEd where the lot generator reports missing
tilesets.