Allow a door and/or a window in the north-west corners of a room.
The north-west corner tile does not support doors or windows, so if a door
or window is placed in a north-west corner of a room, the single NW tile is
split into a W and N tile as appropriate.  If there are already 2 different
wall tiles meeting at a NW corner then that case is handled as well.

To support both a door frame and a window in the same location, a new
layer called "Windows" was added.  Previously, both window and door frame
tiles went into the same "Frames" layer.

It isn't possible to have 2 doors or 2 windows in the same location, that
would require adding a "Doors2" and "Windows2" layers.