Use threads to load tileset images in the background.
When reading a map, MapReader no longer reads the tileset images.
Instead, MapReader creates totally valid tilesets but with all-white tile
pixmaps.  When TilesetManager::addReference() is called, it adds a tileset
to the TilesetImageCache, and then sends that cache-tileset to a job queue
running in a worker thread to load the image.  When a thread finishes
loading an image, TilesetManager sets the tile pixmaps of the cache-tileset
from the image, then updates every tileset using that same image, replacing
the all-white tile pixmaps with the valid image data.  The tilesetChanged()
signal is then emitted to update the display.

When creating a thumbnail image, MapImageManager must wait for tileset images
to be loaded before generating the thumbnail image.