Make sure custom views emit selectionChanged and currentChanged signals.
When a QAbstractItemModel calls reset(), selection-models in views using
that QAbstractItemModel don't emit the selectionChanged or currentChanged
signal.  There were numerous places where this was compensated for by calling
methods to synch the UI when resetting tile, category and furniture lists.

Methods were added that ensure the selectionModel() of a view gets to emit
its signals.