MapManager does all its map loading in threads.
loadMap() uses a thread to do the loading, processing events until the map
is done loading (or fails to load) - unless asynch loading was asked for.

fileChangedTimeout() uses a thread to load a map. mapLoadedByThread() will
do the job of replacing any already-loaded map with the newly-loaded one.
In that case, the mapAboutToChange()/mapChanged() signals are emitted.
mapChanged() is a new signal that other classes were modified to make use of,
rather than mapFileChanged().