Rewrote MiniMap threading to use the "worker object" pattern.
Instead of subclassing QThread, an object of class MiniMapRenderWorker is
created and moved onto a thread of class InterruptibleThread.  Also, instead
of stopping the render thread and updating the minimap's MapComposite in the
application thread every time a change to a map occurs, all changes to a map
are passed as objects of type MapChange to the worker object.

The size of the minimap is no longer tied to the sceneRect of the MapScene
for the map being edited; instead, it is as large as the minimap's MapComposite.
This was an issue before because the sceneRect could change due to layer and
lot visibility, resulting in the minimap not being the right size to display
the fully-visible MapComposite.

InterruptibleThread and BaseWorker were revamped which affected all the other
classes using these objects.