BuildingEd's object tools are much less fidgety now.
The cursor object used to disappear when the mouse pointer was in the center
of a tile or in a corner.  Now the cursor object is always displayed (unless
the coordinates are out-of-bounds) and objects are aligned based on the nearest
edge (or edges).

The furniture tool no longer requires the Shift key to choose between 2 different
orientations in corners, it works much the same as the other tools.