Fixed MapComposite not synching layer opacity correctly.
MapComposite::addMap() was setting the opacity of layers in the new map to
match the opacity of its layers, rather than using the mLayerOpacity value
like it should.  Also, it wasn't doing it recursively on submaps of the submap.

When an already-loaded submap was reloaded, the opacity of layers in that map
were being reset to 1.0.

Another issue with CompositeLayerGroup::synch() was that submaps weren't
using the layer name/opacity/visibility of the root map, only the parent
map.