Door and window objects don't replace the wall unless they align with the wall.
Previously, any time a door or window was placed next to a wall, the wall
tile would be replaced to have a hole in it.  That only makes sense if the
door or window is "in" the wall.  When placing a bathroom stall door next
to a wall there shouldn't be a hole in the adjacent wall.