Use QtSingleApplication to open files in an already-running instance.
Mash wanted a way of opening maps in TileZed from WorldEd.  I found that opening
files using QDesktopServices (or double-clicking a .tmx file in Explorer)
would start a new TileZed instance each time.  So I googled around and settled
on using QtSingleApplication.

When a new instance of TileZed starts, if any files were passed on the command
line, they are passed to another running instance (if any).  If no files are
passed on the command line, the new instance starts up, allowing multiple
instances of TileZed to run.  I did that because the IndieStone peeps sometimes
run 2 copies of TileZed.