Improvements to "Convert To Lot".
Fixed the map-file browse-button not doing anything.
Fixed not being able to add a lot-object the the top-most ObjectGroup.
Allow replacing an existing file.
Added checkbox for choosing Isometric or LevelIsometric orientation.
Added checkbox for opening the new lot in TileZed.

RoomDef objects are now added to the new lot map.