BmpBlender doesn't create floor tiles for black BMP pixels.
For the purposes of blending, if there is a black pixel in the 0/main BMP
that is adjacent to a non-black pixel and the user-drawn tile at that
location matches one of the Rules.txt tiles, we replace the black pixel
with the rule's color.  By pretending that this pixel was painted, we get
the desired "blend" tiles adjacent to non-black pixels and get 0_Vegetation
tiles working when the "fake pixel" matches a 1/veg rule's condition-color.

BmpBlender already did the above, however it placed a tile in the 0_Floor
layer for the "fake" pixel, which was a bit confusing because that pixel
was never painted by the user.  Now the tile is *not* placed in 0_Floor.