Added BuildingPropertiesDialog and building exterior grime.
Rooms already had interior floor- and wall-grime.  But there was no way to
place grime on the outside of buildings except for painting grime tiles by
hand.

Buildings now have a GrimeWall tile.

The BuildingPreferencesDialog allows the user to change the tiles assigned
to the a building (including exterior wall, roof tiles, etc).

Fixed a bug that checked if a room had a GrimeFloor tile when it should have
checked for GrimeWall.