Added command for automatically creating RoomDef objects.
There is a new RoomDefecator submenu in the Tools menu.

The first command will create RoomDef layers/objects in the selected area
or the whole map if nothing is selected.  This uses a flood-fill technique
that separates rooms based on floor tile and hitting walls.  Wall tiles are
identified by using BuildingEd's tiles/furniture.

The second command will merge selected RoomDef objects into a single room.

The third command will delete RoomDef objects on every level in the selected
area, or the whole map if nothing is selected.