Improved the appearance and behavior of TileLayersPanel.
It used to use MixedTilesetView, now it has a custom class.

The layer name is centered horizontally, and is underlined when the mouse
hovers over it to indicate it can be clicked.

The current layer is surrounded by a black rectangle rather than painting
the current highlight color over it, so it is easy to see whether a layer
is visible nor not (since the bg color is gray when a layer is hidden).

Added mouse-over highlighting.