BmpBlender doesn't put a tile in the map if the current map tile is wanted.
The map had a blend tile that should have been there.  The blender then
put another tile in the same layer that the map tile was supposed to be
overriding according to the normal blending rules.

A side effect of this change is that the blender now blends 2 squares adjacent
to where pixels are, rather than 1.