BmpBlender rewrite: Use tile pointers instead of tile names.
Basically, this replaces QStringList with QVector<Tile*> everywhere.
This resulted in a massive speedup of the blending code.
Another benefit is that any unknown tiles or aliases are displayed using the
TilesetManager's "missing" tile.

This did result in a change to which random rule tile is picked.  The old code
picked a random tile/alias, and then if an alias was picked, chose a random
tile from the alias.  The new code keeps a vector of all the tiles used by
a rule (with aliases already resolved) and picks a random tile from that vector.