Improvement to NoBlendTool.
Hide the cursor and prevent painting when the current layer is not a blend
layer.  I tried enabling/disabling the tool when the current layer isn't a
blend layer, but that was annoying due to another tool becoming selected.

Only erase known blend tiles.

Fixed incorrect QRect being used when setting bits.