Improvments to CurbTool and EdgeTool.
Curb.mLayer and Edges.mLayer specify the name of a layer to switch to when
a curb/edge is selected in the tool's dialog.  This replaces the "Layer" field
in CurbToolDialog.

EdgeToolDialog gained an option to suppress blend tiles.

Curbs.txt and Edges.txt are read again if they are modified.  This happens when
switching to the tool.

CurbFile and EdgeFile check for valid block/value names.