Keep separate app/user Curbs.txt/Edges.txt/Fences.txt/LuaTools.txt.
I used to copy these files from the application directory to the user's
~/.TileZed directory if they didn't exist there.  Problem with that was:
how to apply application updates to those files to the user's copy?
Now both the user's file (if one exists) and the application's file are loaded.
This way the user can create their own curbs/edges/fences/tools without fear
of losing them when updating TileZed, and updates made to TileZed are used
as well.