BuildingEd: Fixed hand cursor not displaying after mousing-over a roof's resize-handle.
QGraphicsView sets the cursor (when the mouse is over an item) on the
viewport widget, not on itself.  It seems there is a bug though, because
QWindow::unsetCursor() is never called after leaving an item if the viewport
never had a cursor set on it; QWindow::setCursor() is called with whatever
cursor was set on the viewport widget before mousing over the item.